// car-angular/src/app/clases/mostrarUsuario.interface.ts
export interface mostrarUsuario{
        id: String;
        nombre: String;
        apellido: String;
        tipo_documento: String;
        documento: String;
        lugar_na: String;
        fecha_na: String;
        email: String;
        telefono: String;
        lu_residencia: String;
        sexo: String;
}